+++ ft_strchr.c
@@ -6,7 +6,7 @@
 /*   By: eprusako <eprusako@student.hive.fi>        +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2020/06/05 11:13:58 by eprusako          #+#    #+#             */
-/*   Updated: 2020/06/23 17:49:06 by eprusako         ###   ########.fr       */
+/*   Updated: 2020/06/30 09:35:09 by eprusako         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -14,8 +14,8 @@
 
 char		*ft_strchr(const char *s, int ch)
 {
-	int		i;
-	char	*new;
+	size_t		i;
+	char		*new;
 
 	i = 0;
 	new = (char*)s;
