+++ ft_strncat.c
@@ -6,7 +6,7 @@
 /*   By: eprusako <eprusako@student.hive.fi>        +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2020/06/06 14:10:06 by eprusako          #+#    #+#             */
-/*   Updated: 2020/06/15 16:32:14 by eprusako         ###   ########.fr       */
+/*   Updated: 2020/06/30 09:30:37 by eprusako         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -14,8 +14,8 @@
 
 char		*ft_strncat(char *dst, const char *src, int len)
 {
-	int i;
-	int ii;
+	size_t		i;
+	size_t		ii;
 
 	i = 0;
 	ii = 0;
